requirements.txt
asgiref==3.6.0
colorama==0.4.6
Django==4.2
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
flake8==6.0.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.3.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
